round, rsqrt, sigmoid, isNan, isInf, isFinite, softplus, sqrt, step
