+++ tfjs-core/src/gradients/Abs_grad.ts
@@ -19,7 +19,7 @@ import {Abs} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
 import {cast} from '../ops/cast';
 import {mul} from '../ops/mul';
-import {step} from '../ops/unary_ops';
+import {step} from '../ops/step';
 import {Tensor} from '../tensor';
 
 export const absGradConfig: GradConfig = {
