+++ tfjs-core/src/gradients/Acos_grad.ts
@@ -21,9 +21,9 @@ import {cast} from '../ops/cast';
 import {div} from '../ops/div';
 import {neg} from '../ops/neg';
 import {scalar} from '../ops/scalar';
+import {sqrt} from '../ops/sqrt';
 import {square} from '../ops/square';
 import {sub} from '../ops/sub';
-import {sqrt} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
 
 export const acosGradConfig: GradConfig = {
