+++ tfjs-core/src/gradients/Asinh_grad.ts
@@ -21,8 +21,8 @@ import {add} from '../ops/add';
 import {cast} from '../ops/cast';
 import {div} from '../ops/div';
 import {scalar} from '../ops/scalar';
+import {sqrt} from '../ops/sqrt';
 import {square} from '../ops/square';
-import {sqrt} from '../ops/unary_ops';
 import {Tensor} from '../tensor';
 
 export const asinhGradConfig: GradConfig = {
