+++ tfjs-core/src/gradients/Relu6_grad.ts
@@ -19,7 +19,7 @@ import {GradConfig} from '../kernel_registry';
 import {cast} from '../ops/cast';
 import {lessEqual} from '../ops/less_equal';
 import {mul} from '../ops/mul';
-import {step} from '../ops/unary_ops';
+import {step} from '../ops/step';
 import {Tensor} from '../tensor';
 
 export const relu6GradConfig: GradConfig = {
