+++ tfjs-core/src/gradients/Softplus_grad.ts
@@ -15,21 +15,18 @@
  * =============================================================================
  */
 
-import {Asin} from '../kernel_names';
+import {Softplus} from '../kernel_names';
 import {GradConfig} from '../kernel_registry';
-import {cast} from '../ops/cast';
-import {div} from '../ops/div';
-import {scalar} from '../ops/scalar';
-import {sqrt} from '../ops/sqrt';
-import {square} from '../ops/square';
-import {sub} from '../ops/sub';
+import {mul} from '../ops/mul';
+import {sigmoid} from '../ops/sigmoid';
 import {Tensor} from '../tensor';
 
-export const asinGradConfig: GradConfig = {
-  kernelName: Asin,
+export const softplusGradConfig: GradConfig = {
+  kernelName: Softplus,
   inputsToSave: ['x'],
   gradFunc: (dy: Tensor, saved: Tensor[]) => {
     const [x] = saved;
-    return {x: () => div(dy, sqrt(sub(scalar(1), square(cast(x, 'float32')))))};
+
+    return {x: () => mul(dy, sigmoid(x))};
   }
 };
