+++ tfjs-core/src/ops/ops.ts
@@ -79,6 +79,9 @@ export {gather} from './gather';
 export {greater} from './greater';
 export {greaterEqual} from './greater_equal';
 export {imag} from './imag';
+export {isFinite} from './is_finite';
+export {isInf} from './is_inf';
+export {isNaN} from './is_nan';
 export {leakyRelu} from './leaky_relu';
 export {less} from './less';
 export {lessEqual} from './less_equal';
@@ -138,10 +141,13 @@ export {reverse1d} from './reverse_1d';
 export {reverse2d} from './reverse_2d';
 export {reverse3d} from './reverse_3d';
 export {reverse4d} from './reverse_4d';
+export {round} from './round';
+export {rsqrt} from './rsqrt';
 export {scalar} from './scalar';
 export {selu} from './selu';
 export {separableConv2d} from './separable_conv2d';
 export {setdiff1dAsync} from './setdiff1d_async';
+export {sigmoid} from './sigmoid';
 export {sign} from './sign';
 export {sin} from './sin';
 export {sinh} from './sinh';
@@ -151,12 +157,15 @@ export {slice2d} from './slice2d';
 export {slice3d} from './slice3d';
 export {slice4d} from './slice4d';
 export {softmax} from './softmax';
+export {softplus} from './softplus';
 export {spaceToBatchND} from './space_to_batch_nd';
 export {split} from './split';
+export {sqrt} from './sqrt';
 export {square} from './square';
 export {squaredDifference} from './squared_difference';
 export {squeeze} from './squeeze';
 export {stack} from './stack';
+export {step} from './step';
 export {stridedSlice} from './strided_slice';
 export {sub} from './sub';
 export {sum} from './sum';
@@ -181,7 +190,6 @@ export {zeros} from './zeros';
 export {zerosLike} from './zeros_like';
 
 export * from './boolean_mask';
-export * from './unary_ops';
 export * from './compare';
 export * from './binary_ops';
 export * from './transpose';
