+++ tfjs-core/src/ops/step.ts
@@ -1,0 +1,50 @@
+/**
+ * @license
+ * Copyright 2018 Google LLC. All Rights Reserved.
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ * =============================================================================
+ */
+
+import {ENGINE} from '../engine';
+import {Step, StepAttrs, StepInputs} from '../kernel_names';
+import {NamedAttrMap} from '../kernel_registry';
+import {Tensor} from '../tensor';
+import {NamedTensorMap} from '../tensor_types';
+import {convertToTensor} from '../tensor_util_env';
+import {TensorLike} from '../types';
+
+import {op} from './operation';
+
+/**
+ * Computes step of the input `tf.Tensor` element-wise: `x > 0 ? 1 : alpha * x`
+ *
+ * ```js
+ * const x = tf.tensor1d([0, 2, -1, -3]);
+ *
+ * x.step(.5).print();  // or tf.step(x, .5)
+ * ```
+ * @param x The input tensor.
+ * @param alpha The gradient when input is negative.
+ */
+/** @doc {heading: 'Operations', subheading: 'Basic math'} */
+function step_<T extends Tensor>(x: T|TensorLike, alpha = 0.0): T {
+  const $x = convertToTensor(x, 'x', 'step');
+
+  const inputs: StepInputs = {x: $x};
+  const attrs: StepAttrs = {alpha};
+
+  return ENGINE.runKernelFunc(
+      backend => backend.step($x, alpha), inputs as {} as NamedTensorMap,
+      null /* grad */, Step, attrs as {} as NamedAttrMap);
+}
+export const step = op({step_});
