+++ tfjs-core/src/tests.ts
@@ -113,6 +113,9 @@ import './ops/hann_window_test';
 import './ops/hinge_loss_test';
 import './ops/huber_loss_test';
 import './ops/in_top_k_test';
+import './ops/is_finite_test';
+import './ops/is_inf_test';
+import './ops/is_nan_test';
 import './ops/leaky_relu_test';
 import './ops/less_equal_test';
 import './ops/less_test';
@@ -169,10 +172,13 @@ import './ops/reverse_3d_test';
 import './ops/reverse_4d_test';
 import './ops/reverse_test';
 import './ops/rotate_with_offset_test';
+import './ops/round_test';
+import './ops/rsqrt_test';
 import './ops/scatter_nd_test';
 import './ops/selu_test';
 import './ops/setdiff1d_async_test';
 import './ops/sigmoid_cross_entropy_test';
+import './ops/sigmoid_test';
 import './ops/sign_test';
 import './ops/sin_test';
 import './ops/sinh_test';
@@ -184,11 +190,15 @@ import './ops/slice_test';
 import './ops/slice_util_test';
 import './ops/softmax_cross_entropy_test';
 import './ops/softmax_test';
+import './ops/softplus_test';
 import './ops/space_to_batch_nd_test';
 import './ops/sparse_to_dense_test';
 import './ops/spectral_ops_test';
 import './ops/split_test';
+import './ops/sqrt_test';
+import './ops/square_test';
 import './ops/stack_test';
+import './ops/step_test';
 import './ops/stft_test';
 import './ops/strided_slice_test';
 import './ops/sub_test';
@@ -200,7 +210,6 @@ import './ops/to_pixels_test';
 import './ops/topk_test';
 import './ops/transpose_test';
 import './ops/truncated_normal_test';
-import './ops/unary_ops_test';
 import './ops/unsorted_segment_sum_test';
 import './ops/unstack_test';
 import './ops/where_async_test';
